added BaseQueryResponseTable
