+++ sunpy/net/hek/hek.py
@@ -5,7 +5,6 @@ import json
 import codecs
 import urllib
 import inspect
-import warnings
 from itertools import chain
 
 import astropy.table
@@ -14,10 +13,9 @@ from astropy.time import Time
 
 import sunpy.net._attrs as core_attrs
 from sunpy.net import attr
-from sunpy.net.base_client import BaseClient, BaseQueryResponse
+from sunpy.net.base_client import BaseClient, BaseQueryResponseTable
 from sunpy.net.hek import attrs
 from sunpy.util import dict_keys_same, unique
-from sunpy.util.exceptions import SunpyUserWarning
 from sunpy.util.xml import xml_to_dict
 
 __all__ = ['HEKClient', 'HEKResponse']
@@ -35,10 +33,12 @@ def _freeze(obj):
 
 
 class HEKClient(BaseClient):
-    """ Client to interact with the Heliophysics Event Knowledgebase (HEK).
+    """
+    Provides access to the Heliophysics Event Knowledgebase (HEK).
 
     The HEK stores solar feature and event data generated by algorithms and
-    human observers."""
+    human observers.
+    """
     # FIXME: Expose fields in .attrs with the right types
     # that is, not all StringParamWrapper!
 
@@ -140,49 +140,12 @@ class HEKClient(BaseClient):
         return cls.check_attr_types_in_query(qr, required, optional)
 
 
-class HEKResponse(BaseQueryResponse):
+class HEKResponse(BaseQueryResponseTable):
     """
     A container for data returned from HEK searches.
     """
-    def __init__(self, table=None, client=None):
-        """
-        table : `astropy.table.Table`
-        """
-        super().__init__()
-        self.table = table or astropy.table.Table()
-        self._client = client
-
-    @property
-    def client(self):
-        return self._client
-
-    @client.setter
-    def client(self, client):
-        self._client = client
-
-    @property
-    def blocks(self):
-        return list(self.table.iterrows)
-
-    def __len__(self):
-        if self.table is None:
-            return 0
-        else:
-            return len(self.table)
-
     def __iter__(self):
         return (t for t in self.table)
-
-    def build_table(self):
-        return self.table
-
-    def response_block_properties(self):
-        """
-        Returns a set of class attributes on all the response blocks.
-        """
-        warnings.warn("The HEK client does not support response block properties.",
-                      SunpyUserWarning)
-        return set()
 
     def __getitem__(self, item):
         table_item = self.table.__getitem__(item)
